Break the summary into its own component.
